pyproject.toml
[tool.poetry]
name = "later"
version = "0.1.0"
description = "Service to save anything for later"
authors = ["Pradyumna Chatterjee <chatterjee.pra@gmail.com>"]
readme = "README.md"
packages = [{include = "pre_ml"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.1"
fastapi = "^0.100.0"
uuid = "^1.30"
pydantic-settings = "^2.0.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.11.2"
pydantic = "^2.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^2.17.5"
autopep8 = "^2.0.2"
isort = "^5.12.0"
autoflake = "^2.2.0"
mypy = "^1.4.1"
types-python-jose = "^3.3.4.8"
types-pymysql = "^1.1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
color=true
exclude = '''
    /(
        \.git
        | \._build
        | \.back.out
        | \.build
        | \.coverage
        | \.dist
        | \.hg
        | \.mypy_cache
        | \.tox
        | \.venv
        | ./src/coverage
        | blib2to3
        | tests/data
    )/
'''
include = '\.pyi?$'
line-length = 119


[tool.isort]
color_output = true
combine_as_imports = true
ensure_newline_before_comments = true
force_alphabetical_sort_within_sections = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
lines_between_sections = 1
multi_line_output = 3
profile = "black"
skip = [
    ".coverage",
    "coverage/*",
    "cov.html",
    ".dockerignore",
    ".env",
    ".github",
    ".gitignore",
    ".html",
    ".md",
    ".python-version",
    ".rst",
    ".xml"
]
skip_gitignore = true
skip_glob = [
    "src/repository/migrations/**",
]
src_paths = [
    "src/",
    "tests/",
]
use_parentheses = true

[tool.mypy]
check_untyped_defs = true
color_output = true
error_summary = true
exclude = "(build|data|dist|docs/src|images|logo|logs|output)/"
ignore_missing_imports = true
pretty = true
python_version = "3.11"
strict_optional = true
warn_no_return = true
warn_return_any = false
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
